before installing react-scroll
